fix: example building + add CI step to catch failure of any example code (#581)
Adding CI job to verify that all examples run as expected and are valid. Currently broken examples are added to an exception list and will be fixed later, but at a minimum, the working examples should not be breaking.

Since adding ESM + CJS support in #561 we can move away from a separate build stage for only the examples. Since the examples will be mostly read, or run as one-off, we can point to building and running with `tsx` which simplifies the whole process.